+++ src/components/Pages/MyItems/MyItems/MySingleItem/MySingleItem.js
@@ -18,17 +18,23 @@ const MySingleItem = ({ products, setUpdateConfirm }) => {
         // const uri = `http://localhost:5000/product/${id}`;
         const uri = `https://secret-reaches-38095.herokuapp.com/product/${id}`;
 
-        await axios.delete(uri)
-            .then(async resp => {
-                await resp.data.acknowledged
-                    ?
-                    setUpdateConfirm(prev => !prev)
-                    :
-                    toast("Couldnot Deliver item.");
-            })
-            .catch(error => {
-                toast(error);
-            })
+        if (window.confirm("Do you really want to delete this record?") === true) {
+            await axios.delete(uri)
+                .then(async resp => {
+                    await resp.data.acknowledged
+                        ?
+                        setUpdateConfirm(prev => !prev)
+                        :
+                        toast("Couldnot Delete item.");
+                    resp.data.acknowledged && toast("Record Deleted!");
+
+                })
+                .catch(error => {
+                    toast(error);
+                })
+        } else {
+            toast("You Canceled");
+        }
     }
